+++ apps/backend/src/modules/points-of-interest/points-of-interest.controller.ts
@@ -343,7 +343,6 @@ class PointsOfInterestController extends BaseController {
 		}>,
 	): Promise<APIHandlerResponse<PointsOfInterestResponseDto[]>> {
 		const { query = null } = options;
-
 		const { items } = await this.pointsOfInterestService.findAll(query);
 
 		return {